pkg/dwarf: restructure dwarf packages
