+++ pkg/dwarf/frame/entries.go
