+++ pkg/dwarf/frame/entries_test.go
