+++ pkg/dwarf/frame/parser.go
@@ -7,7 +7,7 @@ import (
 	"bytes"
 	"encoding/binary"
 
-	"github.com/derekparker/delve/dwarf/util"
+	"github.com/derekparker/delve/pkg/dwarf/internal/util"
 )
 
 type parsefunc func(*parseContext) parsefunc
