+++ pkg/dwarf/internal/util/util.go
