+++ pkg/dwarf/internal/util/util_test.go
