+++ proc/eval.go
@@ -11,7 +11,8 @@ import (
 	"go/token"
 	"reflect"
 
-	"github.com/derekparker/delve/dwarf/reader"
+	"github.com/derekparker/delve/pkg/dwarf/reader"
+
 	"golang.org/x/debug/dwarf"
 )
