+++ proc/proc.go
@@ -17,9 +17,9 @@ import (
 
 	"golang.org/x/debug/dwarf"
 
-	"github.com/derekparker/delve/dwarf/frame"
-	"github.com/derekparker/delve/dwarf/line"
-	"github.com/derekparker/delve/dwarf/reader"
+	"github.com/derekparker/delve/pkg/dwarf/frame"
+	"github.com/derekparker/delve/pkg/dwarf/line"
+	"github.com/derekparker/delve/pkg/dwarf/reader"
 )
 
 // Process represents all of the information the debugger
